requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
dnspython==2.0.0
docutils==0.15.2
future==0.18.2
idna==2.9
importlib-metadata==3.1.0
jmespath==0.9.4
Markdown==3.3.3
pycodestyle==2.6.0
PyJWT==1.7.1
PyMySQL==0.9.3
python-whois==0.7.3
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
whois==0.9.7
zipp==3.4.0
